feat: refactor providers and remove unused components
