--- src/app/ClientRouteHandler.tsx
@@ -1,17 +1,0 @@
-"use client";
-
-import { ReactNode } from "react";
-import { usePathname } from "next/navigation";
-import LoadingBar from "@/components/LoadingBar";
-
-export default function ClientRouteHandler({ children }: { children: ReactNode }) {
-  const pathname = usePathname();
-  const showLoading = pathname?.startsWith("/prompts") ?? false;
-
-  return (
-    <>
-      <LoadingBar active={showLoading} pathname={pathname ?? ""} />
-      {children}
-    </>
-  );
-}
