--- src/app/SessionProvider.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { SessionProvider } from "next-auth/react";
-
-export default function ClientProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <SessionProvider>{children}</SessionProvider>;
-}
